requirements.txt
flask>=2.0.0
pyaudio>=0.2.11
numpy>=1.21.0
webrtcvad>=2.0.10

# ElevenLabs dependencies
elevenlabs
python-dotenv
flask-socketio
python-socketio

# LiveKit dependencies
livekit-agents[sarvam,openai,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2

# Testing dependencies
requests

# Image generation dependencies
openai>=0.28.0
Pillow>=8.0.0
